refactor: size_tをuint32_tに置き換え

+++ src/ldpc/LDPCencoding.hpp
@@ -13,11 +13,11 @@ namespace code::LDPC {
 
 template<CheckMatrix T>
 class GenerationMatrix_encoding {
-	static constexpr std::size_t S = T::sourcesize();
-	static constexpr std::size_t C = T::codesize();
-	static constexpr std::size_t Hsize = C-S;
+	static constexpr std::uint32_t S = T::sourcesize();
+	static constexpr std::uint32_t C = T::codesize();
+	static constexpr std::uint32_t Hsize = C-S;
 
-	inline static std::pair<std::vector<std::bitset<S>>,std::vector<std::pair<std::size_t,std::size_t>>> G;//組織符号の生成行列の右側の転置と対応する行置換
+	inline static std::pair<std::vector<std::bitset<S>>,std::vector<std::pair<std::uint32_t,std::uint32_t>>> G;//組織符号の生成行列の右側の転置と対応する行置換
 
 	static void G_init(const T &H);//Gはインスタンスを生成したときに初めて初期化される
 	auto GT_product(const std::bitset<S> &vec) const;//matpos1とvecの積を取る
@@ -44,15 +44,15 @@ template<CheckMatrix T>
 void GenerationMatrix_encoding<T>::G_init(const T &H){
 	//Hをbitset形式に変換
 	std::vector<std::bitset<C>> Hb(Hsize);
-	for(std::size_t i=0; i<Hsize; ++i) for(const auto &j: H[i]) Hb[i].set(j);
+	for(std::uint32_t i=0; i<Hsize; ++i) for(const auto &j: H[i]) Hb[i].set(j);
 
-	// std::vector<std::pair<std::size_t,std::size_t>> Q;
+	// std::vector<std::pair<std::uint32_t,std::uint32_t>> Q;
 	{//Hbを掃き出し、右側を単位行列にする
 		auto Hp = Hb.rbegin();
 		const auto Hend = Hb.rend();
-		for(std::size_t pivot=C-1; Hp!=Hend; --pivot, ++Hp){
+		for(std::uint32_t pivot=C-1; Hp!=Hend; --pivot, ++Hp){
 
-			std::size_t j=pivot;
+			std::uint32_t j=pivot;
 			decltype(Hp) Hi;
 			do{//pivotとその左から列ごとに1を探す
 				Hi=Hp;
@@ -82,10 +82,10 @@ void GenerationMatrix_encoding<T>::G_init(const T &H){
 
 	//単位行列部分は省略して生成行列を作成　Hbの左側の転置=生成行列の右側
 	G.first.resize(Hsize);
-	for(std::size_t i=0; i<Hsize; ++i){
+	for(std::uint32_t i=0; i<Hsize; ++i){
 		auto &Hbi=Hb[i];
 		auto &GTi=G.first[i];
-		for(std::size_t j=0; j<S; ++j) GTi[j] = Hbi[j];
+		for(std::uint32_t j=0; j<S; ++j) GTi[j] = Hbi[j];
 	}
 }
 
@@ -93,7 +93,7 @@ void GenerationMatrix_encoding<T>::G_init(const T &H){
 template<CheckMatrix T>
 auto GenerationMatrix_encoding<T>::GT_product(const std::bitset<S> &vec) const{
 	std::bitset<Hsize> sol;
-	for(std::size_t i=0; i<Hsize; ++i){
+	for(std::uint32_t i=0; i<Hsize; ++i){
 		std::bitset<S> prod = vec&G.first[i];
 		auto num = prod.count();
 		sol.set(i,static_cast<bool>(num&1));
@@ -119,8 +119,8 @@ template<CheckMatrix T>
 auto GenerationMatrix_encoding<T>::systematic_encode(const std::bitset<S> &information) const{
 	std::bitset<C> code;
 	auto parity = GT_product(information);
-	for(std::size_t i=0, iend=S; i<iend; ++i) code[i]=information[i];
-	for(std::size_t i=S, iend=C; i<iend; ++i) code[i]=parity[i-S];
+	for(std::uint32_t i=0, iend=S; i<iend; ++i) code[i]=information[i];
+	for(std::uint32_t i=S, iend=C; i<iend; ++i) code[i]=parity[i-S];
 	return code;
 }
